refactor(macros): tighten body-inference helpers per /simplify review #994
- Memoize `int|float` union in `ClosureTypeFactory::inferArithmetic()`;
  the result is constant — only the operand-numeric gate varied per call.
- Extract `InitializesPsalmConfigSingleton` trait shared by
  `ClosureTypeFactoryTest` and `CachedClosureTypeFactoryTest`; both
  needed the same `Config`-singleton bootstrap and `tearDownAfterClass`
  to avoid leaking state into `NoEnvOutsideConfigHandlerTest`.
- Add `assertBailsForBody()` helper to dedupe the seven bail tests
  that all rebuilt the same `writeTempFile` + `loadClosureFromFile` +
  `assertNull` plumbing inline.
- Drop dead `?? null` coalesces on guaranteed-present array keys in
  `buildClosureType()`.
