+++ tests/Unit/Util/Ast/CachedClosureTypeFactoryTest.php
@@ -8,11 +8,10 @@ use Closure;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
-use Psalm\Config;
-use Psalm\Internal\EventDispatcher;
 use Psalm\LaravelPlugin\Util\Ast\CachedClosureTypeFactory;
 use Psalm\Type\Atomic\TClosure;
 use Psalm\Type\Union;
+use Tests\Psalm\LaravelPlugin\Unit\Util\Ast\Concerns\InitializesPsalmConfigSingleton;
 
 /**
  * Targeted coverage for the memoizing decorator. The stateless build
@@ -33,27 +32,10 @@ use Psalm\Type\Union;
 #[CoversClass(CachedClosureTypeFactory::class)]
 final class CachedClosureTypeFactoryTest extends TestCase
 {
+    use InitializesPsalmConfigSingleton;
+
     /** @var list<string> */
     private array $tempFiles = [];
-
-    public static function setUpBeforeClass(): void
-    {
-        // PR #994 body-flow inference (exercised by the body-flow test below)
-        // calls `TLiteralString::make()` which requires Psalm's singleton
-        // `Config`. Mirror the same minimal-bootstrap pattern as
-        // {@see ClosureTypeFactoryTest::setUpBeforeClass()} so the cached
-        // wrapper's coverage of that path doesn't depend on the un-cached
-        // factory test class running first.
-        $rc = new \ReflectionClass(Config::class);
-        $instance = $rc->newInstanceWithoutConstructor();
-        $rc->getProperty('instance')->setValue(null, $instance);
-        $rc->getProperty('eventDispatcher')->setValue($instance, new EventDispatcher());
-    }
-
-    public static function tearDownAfterClass(): void
-    {
-        (new \ReflectionClass(Config::class))->getProperty('instance')->setValue(null, null);
-    }
 
     #[\Override]
     protected function setUp(): void
